refactor: 이미지 파일 src/image 폴더로 이동
