+++ src/components/MovieImg.ts
@@ -1,5 +1,5 @@
-import noImg from '../../templates/no_image.svg';
 import { IMAGE_URL } from '../config';
+import noImg from '../images/no_image.svg';
 import { Movie } from '../type/movie';
 import { createElementWithAttribute } from '../utils';
